requirements.txt
absl-py==2.1.0
astunparse==1.6.3
backcall==0.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.5
decorator==5.1.1
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.2
gast==0.6.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h5py==3.9.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
iniconfig==2.0.0
ipython==7.34.0
ipython-genutils==0.2.0
ipython-sql==0.5.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml-dtypes==0.2.0
networkx==3.4.2
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682981337988/work
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt_einsum==3.4.0
packaging==24.2
parso==0.8.4
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.24.1
pluggy==1.5.0
prettytable==3.12.0
prompt_toolkit==3.0.48
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
PyWavelets==1.8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scipy==1.11.3
shapely==2.0.6
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
tenacity==9.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tifffile==2024.9.20
tomli==2.2.1
traitlets==5.14.3
trimesh==4.5.3
typing_extensions==4.12.2
urllib3==2.2.3
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
